+++ cmake/sources/ZigSources.txt
@@ -730,6 +730,7 @@ src/install/PackageManager/patchPackage.zig
 src/install/PackageManager/processDependencyList.zig
 src/install/PackageManager/ProgressStrings.zig
 src/install/PackageManager/runTasks.zig
+src/install/PackageManager/security_scanner.zig
 src/install/PackageManager/updatePackageJSONAndInstall.zig
 src/install/PackageManager/UpdateRequest.zig
 src/install/PackageManager/WorkspacePackageJSONCache.zig
