+++ docs/install/security.md
@@ -1,4 +1,4 @@
-# Security Scanners
+# Security Scanner API
 
 Bun's package manager can scan packages for security vulnerabilities before installation, helping protect your applications from supply chain attacks and known vulnerabilities.
