+++ src/install/PackageManager/PackageManagerOptions.zig
@@ -71,8 +71,8 @@ depth: ?usize = null,
 /// isolated installs (pnpm-like) or hoisted installs (yarn-like, original)
 node_linker: NodeLinker = .auto,
 
-// Security provider module path
-security_provider: ?[]const u8 = null,
+// Security scanner module path
+security_scanner: ?[]const u8 = null,
 
 pub const PublishConfig = struct {
     access: ?Access = null,
@@ -282,8 +282,8 @@ pub fn load(
             this.node_linker = node_linker;
         }
 
-        if (config.security_provider) |security_provider| {
-            this.security_provider = security_provider;
+        if (config.security_scanner) |security_scanner| {
+            this.security_scanner = security_scanner;
         }
 
         if (config.cafile) |cafile| {
