+++ src/install/install.zig
@@ -247,7 +247,7 @@ pub const TextLockfile = @import("./lockfile/bun.lock.zig");
 pub const Bin = @import("./bin.zig").Bin;
 pub const FolderResolution = @import("./resolvers/folder_resolver.zig").FolderResolution;
 pub const LifecycleScriptSubprocess = @import("./lifecycle_script_runner.zig").LifecycleScriptSubprocess;
-pub const SecurityScanSubprocess = @import("./PackageManager/install_with_manager.zig").SecurityScanSubprocess;
+pub const SecurityScanSubprocess = @import("./PackageManager/security_scanner.zig").SecurityScanSubprocess;
 pub const PackageInstall = @import("./PackageInstall.zig").PackageInstall;
 pub const Repository = @import("./repository.zig").Repository;
 pub const Resolution = @import("./resolution.zig").Resolution;
